+++ R/rank.functions.R
@@ -35,10 +35,10 @@ rank <- function (x, ...) {
 #' that each agent has a particular rank, for each parameter that has been ranked.
 #'
 #' @examples
+#' \donttest{
 #' # Using the triptans data
 #' network <- mbnma.network(HF2PPITT)
 #'
-#' \donttest{
 #' # Rank predictions from a linear dose-response MBNMA
 #' linear <- mbnma.run(network, fun="linear")
 #' pred <- predict(linear, E0 = 0.5)